Create 92-Special Sum of Array
You have been given an array/list ‘arr’ of length ‘N’, which contains single digit elements at every index.

Your task is to return the sum of all elements of the array. But the final sum should also be a single digit.

To keep the output single digit - you need to keep adding the digits of the output number till a single digit is left.

(https://www.codingninjas.com/studio/problems/special-sum-of-array_893340?leftPanelTab=0)

For Example:

For the given array [5, 8, 4, 9]

The sum of the elements of the array will be 5 + 8 + 4 + 9 = 26.

Since 26 is not a single-digit number, we will again take the sum of the digits of 26. 2 + 6 = 8.

Now 8 is a single-digit number. So we will stop here and return 8.

Input Format:
The first line of input contains a single integer ‘N’, representing the size of the array.

The second line of input contains ‘N’ space-separated integers representing the elements of the given array.

Output Format:
Print a single-digit integer representing the sum of the array.

Example 1:
Input:

5
8 7 0 1 2
Output:

9
Explanation:

For the given array [8, 7, 0, 1, 2]

The sum of the elements of the array will be 8 + 7 + 0 + 1 + 2 = 18.

Since 18 is not a single-digit number, we will again take the sum of the digits of 18. 1 + 8 = 9.

Now 9 is a single-digit number. So we will stop here and return 9.

Example 2:
Input:

4
4 2 1 1
Output:

8
Explanation:

For the given array [4, 2, 1, 1]

The sum of the elements of the array will be 4 + 2+ 1 + 1 = 8.

Since 8 is a single-digit number, we will just return 8.

Constraints:
1 <= N <= 10^3

0 <= arr[i] <= 9