Update and rename 70-Search in Bitonic Array! to 70-Check Sort Order
Problem Statement:
You are given an integer array arr of size n. Your task is to determine if the array is sorted in ascending order, descending order, or if it is not sorted at all. The array may contain duplicate elements. Write a method that returns one of the following:

ASCENDING if the array is sorted in ascending order.
DESCENDING if the array is sorted in descending order.
NOT_SORTED if the array is not sorted.


To simplify your implementation, use the following enum to represent the sort order:

enum SortOrder {
    ASCENDING, DESCENDING, NOT_SORTED
}

Method Signature:
public SortOrder checkSortOrder(int[] arr)

Input:
An integer array arr of size n where  0≤ 𝑛 ≤ 10^5
 
Output:
Returns one of the enum values: ASCENDING, DESCENDING, or NOT_SORTED.

Examples:
1: 
int[] arr = {1, 2, 3, 4, 5};
// Output: ASCENDING

2:
int[] arr = {5, 4, 3, 2, 1};
// Output: DESCENDING

3:
int[] arr = {3, 1, 4, 2, 5};
// Output: NOT_SORTED

4:
int[] arr = {2};
// Output: NOT_SORTED


Constraints:
=>The array may contain duplicate elements.
>The array length 'n' can be zero or more, with a maximum length of 10^5
=>An array of size less than or equal to 1 is considered not sorted.